+++ database/migrations/2020_03_11_164353_create_order_details_table.php
@@ -18,8 +18,8 @@ class CreateOrderDetailsTable extends Migration
             $table->integer('idOrder');
             $table->integer('idProduct');
             $table->integer('quantity');
-            $table->float('price');
-            $table->float('promotion')->default(0)->nullable();
+            $table->integer('price');
+            $table->integer('promotion')->default(0)->nullable();
             $table->timestamps();
         });
     }
